Lint, typecheck
