+++ packages/playground/blueprints/src/lib/steps/run-php.ts
@@ -27,5 +27,7 @@ export const runPHP: StepHandler<RunPHPStep, Promise<PHPResponse>> = async (
 	playground,
 	{ code }
 ) => {
-	return await playground.run({ code });
+	const result = await playground.run({ code });
+	console.log(result.text);
+	return result;
 };
