+++ packages/playground/cli/package.json
@@ -17,7 +17,7 @@
 	],
 	"publishConfig": {
 		"access": "public",
-		"directory": "../../../dist/packages/wp-playground/cli"
+		"directory": "../../../dist/packages/playground/cli"
 	},
 	"license": "GPL-2.0-or-later",
 	"type": "module",
