+++ packages/playground/cli/project.json
@@ -1,14 +1,14 @@
 {
 	"name": "wp-playground-cli",
 	"$schema": "../../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "packages/wp-playground/cli/src",
+	"sourceRoot": "packages/playground/cli/src",
 	"projectType": "library",
 	"targets": {
 		"build": {
 			"executor": "@wp-playground/nx-extensions:package-json",
 			"options": {
-				"tsConfig": "packages/wp-playground/cli/tsconfig.lib.json",
-				"outputPath": "dist/packages/wp-playground/cli",
+				"tsConfig": "packages/playground/cli/tsconfig.lib.json",
+				"outputPath": "dist/packages/playground/cli",
 				"buildTarget": "php-wasm-cli:build:bundle:production"
 			}
 		},
@@ -16,8 +16,8 @@
 			"executor": "@nx/vite:build",
 			"outputs": ["{options.outputPath}"],
 			"options": {
-				"main": "dist/packages/wp-playground/cli/main.js",
-				"outputPath": "dist/packages/wp-playground/cli"
+				"main": "dist/packages/playground/cli/main.js",
+				"outputPath": "dist/packages/playground/cli"
 			},
 			"defaultConfiguration": "production",
 			"configurations": {
@@ -38,7 +38,7 @@
 		"start": {
 			"executor": "@wp-playground/nx-extensions:built-script",
 			"options": {
-				"scriptPath": "dist/packages/wp-playground/cli/main.js"
+				"scriptPath": "dist/packages/playground/cli/main.js"
 			},
 			"dependsOn": ["build"]
 		},
@@ -53,14 +53,14 @@
 			"executor": "@nx/linter:eslint",
 			"outputs": ["{options.outputFile}"],
 			"options": {
-				"lintFilePatterns": ["packages/wp-playground/cli/**/*.ts"]
+				"lintFilePatterns": ["packages/playground/cli/**/*.ts"]
 			}
 		},
 		"test": {
 			"executor": "@nx/jest:jest",
 			"outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
 			"options": {
-				"jestConfig": "packages/wp-playground/cli/jest.config.ts",
+				"jestConfig": "packages/playground/cli/jest.config.ts",
 				"passWithNoTests": true
 			},
 			"configurations": {
@@ -74,8 +74,8 @@
 			"executor": "nx:run-commands",
 			"options": {
 				"commands": [
-					"tsc -p packages/wp-playground/cli/tsconfig.lib.json --noEmit",
-					"tsc -p packages/wp-playground/cli/tsconfig.spec.json --noEmit"
+					"tsc -p packages/playground/cli/tsconfig.lib.json --noEmit",
+					"tsc -p packages/playground/cli/tsconfig.spec.json --noEmit"
 				]
 			}
 		}
